refactor: SRP for modal and tab content
